telegram_parser/requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
